requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.0
cffi==1.17.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.3
boto3>=1.34.0
celery>=5.3.0
redis>=5.0.0
numpy>=1.20.0
numpy-stl>=2.17.0
razorpay>=1.3.0
django-cors-headers>=4.3.1

# python-occ-core>=7.7.0 # Required for STEP/IGES analysis (Install via Conda)